+++ lib/actioncable-next.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module ActionCableNext
-  VERSION = "0.3.3"
+  VERSION = "0.3.4"
 end